// src/lib/enums/product.enum.ts
export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}
export enum ProductVolume {
  // Liquids
  ML_250 = 250,
  ML_330 = 330,
  ML_500 = 500,
  ML_700 = 700,
  ML_1000 = 1000,
  ML_1500 = 1500,
  ML_2000 = 2000,
}
// ======
export enum ProductUnit {
  GRAM = "GRAM",
  KILOGRAM = "KILOGRAM",
  MILLILITER = "MILLILITER",
  LITER = "LITER",
  PIECE = "PIECE",
  CAPSULE = "CAPSULE",
  TABLET = "TABLET",
}
// ========

export enum ProductCollection {
  SUPPLEMENT = "SUPPLEMENT",
  PROTEIN = "PROTEIN",
  VITAMIN = "VITAMIN",
  MINERAL = "MINERAL",
  OTHER = "OTHER",
}

// export enum ProductSize {
//   SMALL = "SMALL",
//   NORMAL = "NORMAL",
//   LARGE = "LARGE",
//   SET = "SET",
// }

// export enum ProductVolume {
//   HALF = 0.5,
//   ONE = 1,
//   ONE_POINT_TWO = 1.2,
//   ONE_POINT_FIVE = 1.5,
//   TWO = 2,
// }

// export enum ProductStatus {
//   PAUSE = "PAUSE",
//   PROCESS = "PROCESS",
//   DELETE = "DELETE",
// }